[MER-3360] Run av stack tidy on av stack sync
I think this should happen automatically on av stack sync. Otherwise, av
stack sync will result in an error that it cannot find a branch in the
repository.
