+++ cmd/av/stack_sync.go
@@ -59,6 +59,10 @@ base branch.
 		if err != nil {
 			return err
 		}
+		if _, err = actions.TidyDB(repo, db); err != nil {
+			return err
+		}
+
 		tx := db.WriteTx()
 		defer tx.Abort()
